requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.12
djangorestframework==3.12.4
drf-yasg2==1.19.4
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pkgutil_resolve_name==1.3.10
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.30.0
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tablib==3.4.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
